# Conflicts:
#	src/modules/exobiology/models/spawn_source.rs
#	src/modules/exobiology/models/spawn_source/target_planet.rs
#	src/modules/exobiology/models/variant.rs
#	src/modules/galaxy/models/planet_class.rs
#	src/modules/logs/blocking/log_dir_reader.rs
#	src/modules/logs/content/log_event_content/codex_entry_event.rs
#	src/modules/state/models/game_state.rs
#	src/modules/state/models/resolvers/planet_state_resolver.rs
#	src/modules/state/models/resolvers/system_state_resolver.rs

+++ src/modules/exobiology/models/genus.rs
@@ -273,7 +273,7 @@ impl Display for Genus {
 
 impl Genus {
     /// The minimum distance in meters required between two samples.
-    pub fn minimum_distance(&self) -> u16 {
+    pub fn minimum_distance(&self) -> u32 {
         match self {
             Genus::Aleoida => 150,
             Genus::AmphoraPlant => 100,
